Created a window and moved serial readline code to a new thread
A new window allows us to start implemented the window to be moved to the tray. It also allows us to easily create a bool flag that allows us to stop all the while loops from running when the app gets closed.

Serial readline is now moved to a new thread, readline blocks any code from running, which causes issues with the window